requirements.txt
langchain
openai
requests
python-dotenv